Update index.js
Changing FAN Powermode:
If set to FAN only Power on/off will only affect the fan (i.e. when device is off and FAN is clicked device powers on in FAN or Dehumification mode (depending on config)
When FAN is clicked to power off but the device is not in FAN/Dehumi mode this click is ignored.
When FANPowerMode is set to complete Device clicking on the FAN will affect the complete device, regardless of the mode. When powering on this way the state of the device is restored.
Some Calls are now called without callback, so the code now checks whether the callback is defined before calling it.
The change to the config.json belongs to this commit (has to be fixed in GIT)

Update config.schema.json

Added new Config setting "fanPowerMode" to select if FAN is independent device or combined with AC. When combined FAN Power on/Off will harm device power on/off.